move tests to edclasses and factories to tests; simplify imports
