+++ src/edclasses/tests/factories/classes_factories.py
@@ -1,6 +1,7 @@
 import factory
-from ..classes import System, Faction, FactionBranch, OrbitalStation
-from ..enums import StationType
+
+from ...classes import System, Faction, FactionBranch, OrbitalStation
+from ...enums import StationType
 
 
 class SystemFactory(factory.Factory):
